requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.0
reportlab==4.0.7
python-multipart==0.0.6
email-validator==2.1.0
cryptography==41.0.7
pandas==2.1.4
openpyxl==3.1.2
pdfplumber==0.10.3
python-docx==1.1.0
apscheduler==3.10.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psycopg2-binary==2.9.9


